Split up schemas from main file
